requirements.txt
Package                                  Version
---------------------------------------- ---------------
annotated-types                          0.7.0
anyio                                    4.8.0
asgiref                                  3.8.1
backoff                                  2.2.1
bcrypt                                   4.2.1
build                                    1.2.2.post1
cachetools                               5.5.2
certifi                                  2025.1.31
charset-normalizer                       3.4.1
chroma-hnswlib                           0.7.6
chromadb                                 0.6.3
click                                    8.1.8
cohere                                   5.14.0
coloredlogs                              15.0.1
Deprecated                               1.2.18
distro                                   1.9.0
durationpy                               0.9
fastapi                                  0.115.8
fastavro                                 1.10.0
filelock                                 3.17.0
flatbuffers                              25.2.10
fsspec                                   2025.2.0
google-auth                              2.38.0
googleapis-common-protos                 1.68.0
grpcio                                   1.70.0
h11                                      0.14.0
httpcore                                 1.0.7
httptools                                0.6.4
httpx                                    0.28.1
httpx-sse                                0.4.0
huggingface-hub                          0.29.1
humanfriendly                            10.0
idna                                     3.10
importlib_metadata                       8.5.0
importlib_resources                      6.5.2
Jinja2                                   3.1.5
jiter                                    0.8.2
joblib                                   1.4.2
kubernetes                               32.0.1
markdown-it-py                           3.0.0
MarkupSafe                               3.0.2
mdurl                                    0.1.2
mmh3                                     5.1.0
monotonic                                1.6
mpmath                                   1.3.0
networkx                                 3.4.2
numpy                                    2.2.3
oauthlib                                 3.2.2
onnxruntime                              1.20.1
openai                                   1.65.4
opentelemetry-api                        1.30.0
opentelemetry-exporter-otlp-proto-common 1.30.0
opentelemetry-exporter-otlp-proto-grpc   1.30.0
opentelemetry-instrumentation            0.51b0
opentelemetry-instrumentation-asgi       0.51b0
opentelemetry-instrumentation-fastapi    0.51b0
opentelemetry-proto                      1.30.0
opentelemetry-sdk                        1.30.0
opentelemetry-semantic-conventions       0.51b0
opentelemetry-util-http                  0.51b0
orjson                                   3.10.15
overrides                                7.7.0
packaging                                24.2
pandas                                   2.2.3
pillow                                   11.1.0
pip                                      24.3.1
posthog                                  3.15.1
protobuf                                 5.29.3
pyasn1                                   0.6.1
pyasn1_modules                           0.4.1
pydantic                                 2.10.6
pydantic_core                            2.27.2
Pygments                                 2.19.1
PyPika                                   0.48.9
pyproject_hooks                          1.2.0
python-dateutil                          2.9.0.post0
python-dotenv                            1.0.1
pytz                                     2025.1
PyYAML                                   6.0.2
regex                                    2024.11.6
requests                                 2.32.3
requests-oauthlib                        2.0.0
rich                                     13.9.4
rsa                                      4.9
safetensors                              0.5.2
scikit-learn                             1.6.1
scipy                                    1.15.2
sentence-transformers                    3.4.1
setuptools                               75.8.1
shellingham                              1.5.4
six                                      1.17.0
sniffio                                  1.3.1
starlette                                0.45.3
sympy                                    1.13.1
tenacity                                 9.0.0
threadpoolctl                            3.5.0
tokenizers                               0.21.0
torch                                    2.6.0
tqdm                                     4.67.1
transformers                             4.49.0
typer                                    0.15.1
types-requests                           2.32.0.20250301
typing_extensions                        4.12.2
tzdata                                   2025.1
urllib3                                  2.3.0
uvicorn                                  0.34.0
uvloop                                   0.21.0
watchfiles                               1.0.4
websocket-client                         1.8.0
websockets                               15.0
wrapt                                    1.17.2
zipp                                     3.21.0
